Adding GMCM support. Fixing font corruption bug due to handling GetData/SetData when texture data is compressed. Improving configuration serialization support. Refactoring some enumerators. #if DEBUGing out presently-unsupported features.

+++ SpriteMaster/Resample/Scalers/EPX/ScalerInterface.cs
@@ -21,7 +21,7 @@ sealed partial class Scaler {
 			hasAlpha: hasAlpha,
 			luminanceWeight: SMConfig.Resample.Common.LuminanceWeight,
 			gammaCorrected: gammaCorrected,
-			equalColorTolerance: SMConfig.Resample.Common.EqualColorTolerance,
+			equalColorTolerance: (uint)SMConfig.Resample.Common.EqualColorTolerance,
 			useRedmean: SMConfig.Resample.UseRedmean
 		);
 	}
