+++ Tools/xBRZ/PreviewProgram.cs
@@ -12,7 +12,7 @@ static class PreviewProgram {
 	internal static bool AlphaPremultiplication = true;
 	internal static uint Scale = 1;
 	internal static double LuminanceWeight = Config.Resample.Common.LuminanceWeight;
-	internal static uint EqualColorTolerance = Config.Resample.Common.EqualColorTolerance;
+	internal static uint EqualColorTolerance = (uint)Config.Resample.Common.EqualColorTolerance;
 	internal static double DominantDirectionThreshold = Config.Resample.xBRZ.DominantDirectionThreshold;
 	internal static double SteepDirectionThreshold = Config.Resample.xBRZ.SteepDirectionThreshold;
 	internal static double CenterDirectionBias = Config.Resample.xBRZ.CenterDirectionBias;
